Barrel files to the rescue!
