+++ src/modules/auth/auth.resolver.ts
@@ -1,15 +1,14 @@
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
 import { AuthService } from './auth.service';
-import { LoginDTO } from './dtos/login.dto';
-import { LoginRO } from '@modules/auth/dtos/login.ro';
-import { User } from '@modules/users/interfaces/user.interface';
+import { LoginIO, LoginRO } from './dto';
+import { User } from '@modules/users/interfaces';
 
 @Resolver()
 export class AuthResolver {
     constructor(private readonly authService: AuthService) {}
 
     @Mutation(returns => LoginRO)
-    async login(@Args('loginData') payload: LoginDTO): Promise<LoginRO> {
+    async login(@Args('loginData') payload: LoginIO): Promise<LoginRO> {
         const user: User = { ...payload };
         return await this.authService.login(user);
     }
