+++ src/modules/auth/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './login.io';
+export * from './login.ro';
