+++ src/modules/auth/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './jwt-payload.interface';
