+++ src/modules/roles/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './role.ro';
