+++ src/modules/roles/entities/index.ts
@@ -1,5 +1,1 @@
-import { Role } from './role.entity';
-
-export {
-    Role
-};
+export * from './role.entity';
