+++ src/modules/roles/guards/index.ts
@@ -1,0 +1,1 @@
+export * from './roles.guard';
