+++ src/modules/roles/guards/roles.guard.ts
@@ -1,4 +1,4 @@
-import { ExecutionContext, Injectable, UnauthorizedException, CanActivate } from '@nestjs/common';
+import { ExecutionContext, Injectable, CanActivate } from '@nestjs/common';
 import { GqlExecutionContext } from '@nestjs/graphql';
 import { Reflector } from '@nestjs/core';
