+++ src/modules/users/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './createUser.io';
+export * from './updateUser.io';
+export * from './user.ro';
