+++ src/modules/users/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './user.interface';
+export * from './updateUser.interface';
