+++ src/modules/users/interfaces/updateUser.interface.ts
@@ -1,0 +1,5 @@
+export interface UpdateUser {
+    firstName?: string;
+    surname?: string;
+    email?: string;
+}
