+++ src/modules/users/interfaces/user.interface.ts
@@ -5,9 +5,3 @@ export interface User {
     email: string;
     password: string;
 }
-
-export interface UpdateUser {
-    firstName?: string;
-    surname?: string;
-    email?: string;
-}
