+++ test/unit/roles/roles.guard.test.ts
@@ -12,11 +12,11 @@ import { AuthResolver } from '@modules/auth/auth.resolver';
 import { UsersResolver } from '@modules/users/users.resolver';
 import { UsersService } from '@modules/users/users.service';
 import { AuthService } from '@modules/auth/auth.service';
-import { GqlAuthGuard } from '@modules/auth/guards/auth.guard';
-import { RolesGuard } from '@modules/roles/guards/roles.guard';
+import { GqlAuthGuard } from '@modules/auth/guards';
+import { RolesGuard } from '@modules/roles/guards';
 import { Roles } from '@modules/roles/roles.decorator';
 import { User } from '@modules/users/entities';
-import { User as UserInterface } from '@modules/users/interfaces/user.interface';
+import { User as UserInterface } from '@modules/users/interfaces';
 import { constants } from '@utils/helpers/roles.helper';
 import { makeGQLHelperMethods } from '@test/helpers';
