+++ test/unit/users/users.dto.test.ts
@@ -2,10 +2,10 @@ import 'reflect-metadata';
 import * as _ from 'lodash';
 import { validateOrReject } from 'class-validator';
 
-import { CreateUserDTO, UpdateUserDTO } from '@modules/users/users.dto';
+import { CreateUserIO, UpdateUserIO } from '@modules/users/dto';
 
 describe('User (input) DTOs', () => {
-    describe('CreateUserDTO', () => {
+    describe('CreateUserIO', () => {
         const passingObject = {
             firstName: 'John',
             surname: 'Wick',
@@ -16,11 +16,11 @@ describe('User (input) DTOs', () => {
         describe('firstName', () => {
             it('must be longer than or equal 1 char', async () => {
                 // test both ends of the constraint
-                const valid = new CreateUserDTO({
+                const valid = new CreateUserIO({
                     ...passingObject,
                     firstName: 'a'
                 });
-                const invalid = new CreateUserDTO({
+                const invalid = new CreateUserIO({
                     ...passingObject,
                     firstName: ''
                 });
@@ -29,11 +29,11 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(invalid)).rejects.toHaveLength(1);
             });
             it('must be shorter than or equal 100 chars', async () => {
-                const valid = new CreateUserDTO({
+                const valid = new CreateUserIO({
                     ...passingObject,
                     firstName: _.repeat('a', 100)
                 });
-                const invalid = new CreateUserDTO({
+                const invalid = new CreateUserIO({
                     ...passingObject,
                     firstName: _.repeat('a', 101)
                 });
@@ -44,11 +44,11 @@ describe('User (input) DTOs', () => {
         });
         describe('surname', () => {
             it('must be longer than or equal 1 char', async () => {
-                const valid = new CreateUserDTO({
+                const valid = new CreateUserIO({
                     ...passingObject,
                     surname: 'a'
                 });
-                const invalid = new CreateUserDTO({
+                const invalid = new CreateUserIO({
                     ...passingObject,
                     surname: ''
                 });
@@ -57,11 +57,11 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(invalid)).rejects.toHaveLength(1);
             });
             it('must be shorter than or equal 100 chars', async () => {
-                const valid = new CreateUserDTO({
+                const valid = new CreateUserIO({
                     ...passingObject,
                     surname: _.repeat('a', 100)
                 });
-                const invalid = new CreateUserDTO({
+                const invalid = new CreateUserIO({
                     ...passingObject,
                     surname: _.repeat('a', 101)
                 });
@@ -72,8 +72,8 @@ describe('User (input) DTOs', () => {
         });
         describe('email', () => {
             it('can\'t be empty', async () => {
-                const valid = new CreateUserDTO(passingObject);
-                const invalid = new CreateUserDTO({
+                const valid = new CreateUserIO(passingObject);
+                const invalid = new CreateUserIO({
                     ...passingObject,
                     email: ''
                 });
@@ -82,8 +82,8 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(invalid)).rejects.toHaveLength(1);
             });
             it('must be an email', async () => {
-                const valid = new CreateUserDTO(passingObject);
-                const invalid = new CreateUserDTO({
+                const valid = new CreateUserIO(passingObject);
+                const invalid = new CreateUserIO({
                     ...passingObject,
                     email: 'thisIsNotAnEmail'
                 });
@@ -94,11 +94,11 @@ describe('User (input) DTOs', () => {
         });
         describe('password', () => {
             it('must be longer than or equal 6 chars', async () => {
-                const valid = new CreateUserDTO({
+                const valid = new CreateUserIO({
                     ...passingObject,
                     password: _.repeat('a', 6)
                 });
-                const invalid = new CreateUserDTO({
+                const invalid = new CreateUserIO({
                     ...passingObject,
                     password: _.repeat('a', 5)
                 });
@@ -107,11 +107,11 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(invalid)).rejects.toHaveLength(1);
             });
             it('must be shorter than or equal 50 chars', async () => {
-                const valid = new CreateUserDTO({
+                const valid = new CreateUserIO({
                     ...passingObject,
                     password: _.repeat('a', 50)
                 });
-                const invalid = new CreateUserDTO({
+                const invalid = new CreateUserIO({
                     ...passingObject,
                     password: _.repeat('a', 51)
                 });
@@ -121,7 +121,7 @@ describe('User (input) DTOs', () => {
             });
         });
     });
-    describe('UpdateUserDTO', () => {
+    describe('UpdateUserIO', () => {
         const passingObject = {
             firstName: 'John',
             surname: 'Wick',
@@ -130,7 +130,7 @@ describe('User (input) DTOs', () => {
 
         describe('firstName', () => {
             it('can be optional', async () => {
-                const valid = new UpdateUserDTO({
+                const valid = new UpdateUserIO({
                     surname: passingObject.surname,
                     email: passingObject.email
                 });
@@ -138,11 +138,11 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(valid)).resolves;
             });
             it('must be longer than or equal 1 char', async () => {
-                const valid = new UpdateUserDTO({
+                const valid = new UpdateUserIO({
                     ...passingObject,
                     firstName: 'a'
                 });
-                const invalid = new UpdateUserDTO({
+                const invalid = new UpdateUserIO({
                     ...passingObject,
                     firstName: ''
                 });
@@ -151,11 +151,11 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(invalid)).rejects.toHaveLength(1);
             });
             it('must be shorter than or equal 100 chars', async () => {
-                const valid = new UpdateUserDTO({
+                const valid = new UpdateUserIO({
                     ...passingObject,
                     firstName: _.repeat('a', 100)
                 });
-                const invalid = new UpdateUserDTO({
+                const invalid = new UpdateUserIO({
                     ...passingObject,
                     firstName: _.repeat('a', 101)
                 });
@@ -166,7 +166,7 @@ describe('User (input) DTOs', () => {
         });
         describe('surname', () => {
             it('can be optional', async () => {
-                const valid = new UpdateUserDTO({
+                const valid = new UpdateUserIO({
                     firstName: passingObject.firstName,
                     email: passingObject.email
                 });
@@ -174,11 +174,11 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(valid)).resolves;
             });
             it('must be longer than or equal 1 char', async () => {
-                const valid = new UpdateUserDTO({
+                const valid = new UpdateUserIO({
                     ...passingObject,
                     surname: 'a'
                 });
-                const invalid = new UpdateUserDTO({
+                const invalid = new UpdateUserIO({
                     ...passingObject,
                     surname: ''
                 });
@@ -187,11 +187,11 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(invalid)).rejects.toHaveLength(1);
             });
             it('must be shorter than or equal 100 chars', async () => {
-                const valid = new UpdateUserDTO({
+                const valid = new UpdateUserIO({
                     ...passingObject,
                     surname: _.repeat('a', 100)
                 });
-                const invalid = new UpdateUserDTO({
+                const invalid = new UpdateUserIO({
                     ...passingObject,
                     surname: _.repeat('a', 101)
                 });
@@ -202,7 +202,7 @@ describe('User (input) DTOs', () => {
         });
         describe('email', () => {
             it('can be optional', async () => {
-                const valid = new UpdateUserDTO({
+                const valid = new UpdateUserIO({
                     firstName: passingObject.firstName,
                     surname: passingObject.surname
                 });
@@ -210,8 +210,8 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(valid)).resolves;
             });
             it('can\'t be empty', async () => {
-                const valid = new UpdateUserDTO(passingObject);
-                const invalid = new UpdateUserDTO({
+                const valid = new UpdateUserIO(passingObject);
+                const invalid = new UpdateUserIO({
                     ...passingObject,
                     email: ''
                 });
@@ -220,8 +220,8 @@ describe('User (input) DTOs', () => {
                 await expect(validateOrReject(invalid)).rejects.toHaveLength(1);
             });
             it('must be an email', async () => {
-                const valid = new UpdateUserDTO(passingObject);
-                const invalid = new UpdateUserDTO({
+                const valid = new UpdateUserIO(passingObject);
+                const invalid = new UpdateUserIO({
                     ...passingObject,
                     email: 'thisIsNotAnEmail'
                 });
